+++ ui/components/data-sources/index.js
@@ -1,3 +1,4 @@
 export * from "./AddDataSourceDialog";
 export * from "./DataSourcesContainer";
 export * from "./DataSourcesList";
+export * from "./EmptyList";
